Temporarily add modified KaTeX fonts
At the time of writing, the KaTeX font distribution has a "bug" in the
hbar glyph U+210F causing it to appear too heavy.  I regenerated the
files using the main katex-fonts repository as of
katex/katex-fonts@98afa42 and found the issue to be fixed but not yet
pushed to the repository or distributed on the CDNs.

This adds the generated font files to the website until a new KaTeX
distribution becomes available.