Write own syntax highlighter
I still use Pygments for lexing, but I completely rewrote the HTML
formatter and the Markdown extension handling grabbing code blocks.
This gives me much more control over the generated HTML, which let me
style it much more easily and correctly, and we can modify the lexers in
use on-the-fly, and so on.

This fixes the bugs where the line number background colour wouldn't go
all the way to the top and bottom, and removes the extra line breaks
when copying from a code block.

I also want to add a special "console usage" option to the highlighter
extension so I can get nicer output when showing command-line stuff.
This is most useful for Bash bits, but may be useful in the standard
Python interpreter (and maybe even gdb?).