Insert KaTeX MathML nodes raw, not as escaped XML
Previously, since the KaTeX MathML nodes were XML but not HTML, the
Markdown parser was escaping the tags in them.  This didn't show up in
regular webpages, but it destroyed the metadata around equations and
made articles unreadable in reader mode on Firefox.

This fixes it by adding this part of the KaTeX output to the raw HTML
stash in python-markdown, and restoring it only after the HTML escaper
has already run.